requirements.txt
openai==0.26.5
python-dotenv==1.0.0
pyttsx3==2.90
SpeechRecognition==3.9.0
